JSON: Fix MSVC2013 warnings
json_p.h(283) : warning C4800: 'uint64_t' : forcing value to bool 'true' or 'false' (performance warning)
json_p.h(546) : warning C4800: 'uint32_t' : forcing value to bool 'true' or 'false' (performance warning)
qjsonvalue.cpp(176) : warning C4244: '=' : conversion from 'int64_t' to 'double', possible loss of data
qjsonvalue.cpp(518) : warning C4244: 'return' : conversion from 'const double' to 'int', possible loss of data

+++ src/corelib/json/qjsonvalue.cpp
@@ -173,7 +173,7 @@ QJsonValue::QJsonValue(int n)
 QJsonValue::QJsonValue(qint64 n)
     : d(0), t(Double)
 {
-    this->dbl = n;
+    this->dbl = double(n);
 }
 
 /*!
@@ -515,7 +515,7 @@ bool QJsonValue::toBool(bool defaultValue) const
 int QJsonValue::toInt(int defaultValue) const
 {
     if (t == Double && int(dbl) == dbl)
-        return dbl;
+        return int(dbl);
     return defaultValue;
 }
